+++ ios/wrappers/location/NativeLocation.mm
@@ -1,5 +1,5 @@
 #import <React/RCTBridgeModule.h>
-#import <RNCustomerIOSpec/RNCustomerIOSpec.h>
+#import <RNZixflowSpec/RNZixflowSpec.h>
 
 // Objective-C wrapper for new architecture TurboModule implementation
 @interface RCTNativeCustomerIOLocation : NSObject <NativeCustomerIOLocationSpec>